+++ CHANGELOG.md
@@ -11,6 +11,7 @@ v5.7.0
 **IncNavierStokesSolver**
 - Fix initial and boundary conditions in the moving reference frame (!1692)
 - Fix memory-leak for the Mixed_CG_Discontinuous projection when initializing the traceMep (!1806)
+- Add synthetic turbulence generation for the incompressible solver (!1664) 
 
 **ShallowWaterSolver**
 - Implement implicit time-discritization (!1784)